backend/requirements.txt
Django==5.0.3  # Using latest stable 5.0.3, acting as the foundation for the "6.0" features requested
PyMySQL==1.1.0  # Pure-Python alternative to mysqlclient to avoid C++ Build Tools error on Windows
argon2-cffi==23.1.0
python-dotenv==1.0.1
httpx==0.27.0
# spacy==3.7.4  # Commented out temporarily as it can take a very long time to build on Windows without pre-built wheels.
vaderSentiment==3.3.2
google-generativeai==0.4.1
uvicorn==0.28.0
django-cors-headers==4.3.1
PyJWT==2.8.0
redis==5.0.3
django-redis==5.4.0
django-ratelimit==4.1.0
groq==0.5.0

